Move most of PartitionedOutputOperator tests to TestPagePartitioner
